+++ sdk/data/azcosmos/cosmos_partition_key_range_request_options_test.go
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package azcosmos
+
+import (
+	"testing"
+)
+
+func TestPartitionKeyRangeRequestOptionsToHeaders(t *testing.T) {
+	options := &partitionKeyRangeOptions{}
+	if options.toHeaders() != nil {
+		t.Error("toHeaders should return nil")
+	}
+}