+++ sdk/data/azcosmos/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-* Added support for BypassIntegratedCache option
+* Added support for BypassIntegratedCache option See [PR 24772](https://github.com/Azure/azure-sdk-for-go/pull/24772)
 
 ### Breaking Changes