+++ sdk/data/azcosmos/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 * Added support for BypassIntegratedCache option See [PR 24772](https://github.com/Azure/azure-sdk-for-go/pull/24772)
+* Added support for GetPartitionKeyRange from Cosmos DB Container See [PR 24835](https://github.com/Azure/azure-sdk-for-go/pull/24835)
 
 ### Breaking Changes