+++ sdk/data/azcosmos/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added support for specifying Full-Text Search indexing policies when creating a container. See [PR 24833](https://github.com/Azure/azure-sdk-for-go/pull/24833)
 * Added support for specifying Vector Search indexing policies when creating a container. See [PR 24833](https://github.com/Azure/azure-sdk-for-go/pull/24833)
 
+
 ### Breaking Changes
 
 ### Bugs Fixed